+++ test/main.f90
@@ -2,7 +2,9 @@ program main
     use CODATA_constants, only: c=>speed_of_light_in_vacuum, Avogadro_constant, &
         & Boltzmann_constant, electron_volt
     use test
+    use functions
     implicit none
+    character(25) :: s
 
     print *, "Do you know this constant with so many zeros?"
     print '(b64.64)', c
@@ -15,4 +17,10 @@ program main
     print *, "********* all constants ************"
     call tests
 
+    s = "4.5003878060E-10         "
+    call insert_space(s, 6)
+    print *, s
+    s = "2808.92113298            "
+    call insert_space(s, 25)
+    print *, s
 end program main